선물 리스트를 로컬에 담기

+++ app/src/main/java/com/nexters/yetda/android/database/model/Present.kt
@@ -5,13 +5,10 @@ import io.realm.RealmObject
 import io.realm.annotations.PrimaryKey
 import java.util.*
 
-
 open class Present : RealmObject() {
     @PrimaryKey
     var id: Int = 0
-    var price: Int = 0
+    var price: Long = 0
     var name: String = ""
     var tags = RealmList<Tag>()
-
-
 }
